Added first example: monitor new dexscreener listings on Solana.
This example showcases how to convert a feed from a Telegram channel to
a structured feed, that could be used for further data analysis.

It is important because this data is not accessible onchain(these are
listings, not new pairs), and it could be used to determine whether a
"project" is investing, or, determine its author(s) on-chain addresse(s)
by searching for recent dexscreener listing payment, and see if they are
correlated to other launches.